Split devtools into optional package
