+++ Directory.Build.targets
@@ -11,14 +11,14 @@
 
   <!-- Promote IL trimming/AOT warnings to errors on the core Reactor library.
        This ensures new reflection usage must be annotated before it can merge. -->
-  <PropertyGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net10.0')) And '$(MSBuildProjectName)' == 'Reactor'">
+  <PropertyGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net10.0')) And ('$(MSBuildProjectName)' == 'Reactor' Or '$(MSBuildProjectName)' == 'Reactor.Devtools')">
     <WarningsAsErrors>$(WarningsAsErrors);IL2026;IL2062;IL2065;IL2067;IL2069;IL2070;IL2072;IL2073;IL2075;IL2087;IL2090;IL2091;IL2111;IL3000;IL3050</WarningsAsErrors>
   </PropertyGroup>
 
   <!-- Suppress trimming/AOT warnings in consumer projects (tests, samples, CLI).
        The core Reactor library is already warning-free; these projects intentionally
        call reflection-based APIs and will not be AOT-published. -->
-  <PropertyGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net10.0')) And '$(MSBuildProjectName)' != 'Reactor'">
+  <PropertyGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net10.0')) And '$(MSBuildProjectName)' != 'Reactor' And '$(MSBuildProjectName)' != 'Reactor.Devtools'">
     <NoWarn>$(NoWarn);IL2026;IL2062;IL2065;IL2067;IL2070;IL2072;IL2075;IL2090;IL2091;IL3000;IL3050</NoWarn>
   </PropertyGroup>
