+++ samples/Directory.Build.props
@@ -26,10 +26,11 @@
     <ApplicationIcon Condition="'$(ApplicationIcon)' == ''">$(MSBuildThisFileDirectory)Assets\WinUI.ico</ApplicationIcon>
   </PropertyGroup>
 
-  <ItemGroup Condition="'$(Configuration)' == 'Debug' And '$(ReactorDevtoolsManaged)' != 'false' And !$([System.String]::Copy('$(MSBuildProjectDirectory)').Contains('\samples\apps\hello-world-aot'))">
+  <ItemGroup Condition="'$(Configuration)' == 'Debug' And '$(ReactorDevtoolsManaged)' != 'false' And $([System.String]::Copy('$(TargetFramework)').Contains('-windows')) And !$([System.String]::Copy('$(MSBuildProjectDirectory)').Contains('\samples\apps\hello-world-aot'))">
     <RuntimeHostConfigurationOption Include="Reactor.DevtoolsSupport"
                                     Value="true"
                                     Trim="true" />
+    <ProjectReference Include="$(MSBuildThisFileDirectory)..\src\Reactor.Devtools\Reactor.Devtools.csproj" />
   </ItemGroup>
 
 </Project>
