+++ src/Reactor.Devtools/DevtoolsJsonContext.cs
