+++ src/Reactor.Devtools/LockfileRegistry.cs
