+++ src/Reactor.Devtools/LogCaptureBuffer.cs
