+++ src/Reactor.Devtools/LogCaptureInstall.cs
