+++ src/Reactor.Devtools/McpToolRegistry.cs
