+++ src/Reactor.Devtools/NodeIdBuilder.cs
