+++ src/Reactor.Devtools/NodeRegistry.cs
