+++ src/Reactor.Devtools/ScreenshotCapture.cs
