+++ src/Reactor.Devtools/SelectorParser.cs
