+++ src/Reactor.Devtools/SelectorResolver.cs
