+++ src/Reactor.Devtools/TreeWalker.cs
