+++ src/Reactor.Devtools/WindowIdAllocator.cs
