+++ src/Reactor.Devtools/WindowRegistry.cs
